+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Add `CompareOptions` parameter to `LookupTable` constructor allowing customization of how keys are compared. 
 - Add `TryGetPreviousTab` and `TryGetNextTab` functions to `TabGroup`
 - Add `OverflowScrolling` public get-only property to `TabGroup`
+- Add `ActionInvoked` event to `ActionEmitter`
 
 ### Changed
 
@@ -29,6 +30,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - ⚠️Breaking: Rename `GameFile` events to no longer begin with `On` term
 - Improved documentation of many public methods (especially `GameFile`)
 - ⚠️Breaking: Adjust the accessibility of many UI functions and fields
+- ⚠️Breaking: `ActionEmitter` no longer invokes `ActionsUpdated` event when an `ActionEvent` is invoked (use newly created `ActionInvoked` event if relevant)
 
 ## [v2.1.0] - 2024-06-25